move text-layout.scss to the respective component's folder because it generates local selectors only
