+++ epam-promo/helpers/text-layout.scss
